Rename ov_numeric to openvino_numeric
