+++ docs/api/source/conf.py
@@ -145,7 +145,7 @@ mock_modules = [
     "nncf.tensor.functions.torch_linalg",
     "nncf.tensor.functions.torch_io",
     "nncf.tensor.functions.numpy_io",
-    "nncf.tensor.functions.ov_numeric",
+    "nncf.tensor.functions.openvino_numeric",
 ]
 
 with mock(mock_modules):
